wip - refactoring and implementing 'real-time' log output
